# Improve worst case performance when using ARM Neon instructions