… ext/oj/dump.c
… test/data/activitypub.compact.txt
… test/data/activitypub.compat.txt
… test/data/activitypub.json
… test/data/activitypub.rails.txt
… test/data/canada.compat.txt
… test/data/canada.json
… test/data/canada.rails.txt
… test/data/citm_catalog.compat.txt
… test/data/citm_catalog.json
… test/data/citm_catalog.rails.txt
… test/data/ohai.compat.txt
… test/data/ohai.json
… test/data/ohai.rails.txt
… test/data/twitter.compat.txt
… test/data/twitter.json
… test/data/twitter.rails.txt