@ohler55 2025-06-01
There seems to be some confusion with the data files. I noticed some were used in the benchmarks and some were not. I also not that the benchmarks list other files not in the date such as "bytes.15.worstcase".

There is a duplicate data file activitypub.compact.txt and activitypub.compat.txt.

If the files are JSON and they appear to be then why give them a ".json" suffix instead of ".txt"?

Thee other changes tentatively look fine but I'd like to get a better feel for what type of files are slower and which are faster.