@samyron 2025-06-01
> There seems to be some confusion with the data files. I noticed some were used in the benchmarks and some were not. I also not that the benchmarks list other files not in the date such as "bytes.15.worstcase".

The `bytes.15.worstcase` isn't a file, it's just a literal string. These are the synthetic benchmarks included in the results above:

```
benchmark_encoding "bytes.15.bestcase", ([("a" * 15)] * 10000)
benchmark_encoding "bytes.15.worstcase", ([('"' * 15)] * 10000)
benchmark_encoding "bytes.15.worstcase-2", (["\u0001a\u0002b\u0001a\u0002b\u0001a\u0002b\u0001a\u0002"] * 10000)
benchmark_encoding "bytes.32.worstcase", ([("\"\t\\\n" * 8)] * 10000)
```

> There is a duplicate data file activitypub.compact.txt and activitypub.compat.txt.

Oops... apologies about that. I generated the first by hand then wrote a script... but I guess I had a typo in that first one. I will also add a snippet for how these files were generated to hopefully save someone a few minutes in the future should they ever need to be re-generated.

> If the files are JSON and they appear to be then why give them a ".json" suffix instead of ".txt"?

They are, I'll fix the extension.

> Thee other changes tentatively look fine but I'd like to get a better feel for what type of files are slower and which are faster.

Data that is very heavy in bytes that need to be escaped will benefit the most from these changes. Should someone decide to JSON encode binary data, I would expect these changes would help a lot as compared to the current ARM Neon implementation.

Text-based input likely wouldn't see much change from these changes.